EVOLUTION: Turn on the selective update for refined blocks.
Now, the selective update after refinement/derefinement process is on.
Before checking the refinement and update the mesh, We set all blocks to
not be updated. Then, only blocks which are refined or derefined
including all their neighbors, are set to be updated. This reduces the
repeated work, i.e. boundary update and variable conversion, on blocks
which didn't actually change.

After the mesh refinement, boundary and variable update, we mark all
blocks to be updated, so the regular temporal integration process works.

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>