EVOLUTION: Update primitive variables before boundaries.
The new order is to first, convert the updated conserved variables to
their primitive representation in the block interiors, update boundaries
of the primitive variables, and finally, convert the primitive variables
at the ghost cells to conserved ones.

The change updates the primitive variables in the block interiors before
the boundary update.

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>