IO: Make the snapshot number continuously increasing.
If we restart a job with a different snapshot interval, the snapshot
number will continue from a different value corresponding to the
simulation time.

The snapshot number shouldn't be related to the current simulation time.
It should be always increasing.  The simulation time is stored as an
attribute, anyway.  This patch makes the snapshot number always
increasing continuously, without any gaps for restarted jobs.

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>