BLOCKS: Rewrite and micro optimize iterate_over_neighbors().
We prepare a set of indices for all faces in the block.  Those indices
point to edges and corners for each face, and are set one during the
first execution.  Then, we simply iterate over faces and call
iterate_over_face() in order to apply subroutine pprocedure to edges and
corners connected to a given face.

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>