EQUATIONS: Correct negative pressure if desired.
If option "fix_positivity" is on, we check if the pressure obtained from
conservative variables is positive. If it is negative, it is replaced by
a value corresponding to the maximum allowed sonic Mach number, provided
by option "msmax".  The new value of pressure and corresponding total
energy is set in all unphysical cells then.

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>