MESH: Move child processing from update_mesh() to prepare_sibling_derefinement().
After checking refinement flags and process neighbors, we need to check
if all siblings of blocks which were selected for derefinement, are
eligible for derefinement too, i.e. lay at the same level, and are also
selected for derefinement. After this step, we bring all siblings to the
same processor, so the derefinement can be done efficiently.

In this commit, the above step is moved from update_mesh() to
prepare_sibling_derefinement().

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>