BLOCKS: Add face, edge and corner neighbor fields to meta block structure.
This commit extends the meta block structure by fields in which we store
pointers to face, edge and corner neighbors.

In the current implementation the boundary update is only done for block
faces, meaning that the corners are not properly updated between the blocks
on different levels. This, of course, can create unexpected problems,
like improper interpolation in the ghost zones, resulting in NaNs there.

The fields for neighbor pointers should allow for a quick access to neighbors
for full, properly done, boundary update.

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>