BLOCKS: Separate children and neighbor pointer update in refine_block().
In order to avoid inconsistencies in the face, edge and corner neighbor
pointer update in refine_block(), like updating the neighbour's pointer
in the middle while it might be still needed in another pointer update,
separate the children and neighbor pointers update.

First, update the face, edge and corner neighbor pointers for the newly
created children only, and when this step is done, update the neighbor'
face, edge and corner pointers.

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>