IO: Format change for storing data blocks in restart files.
This format change is required to remove unnecessary allocation of large
arrays in order to store primitive and conservative variables in the
restart files. With this format change the memory limit of the half of
the available total memory is removed. However, in the case of use of
the data compression the sizes of restart files are typically large. The
compression ratio, however, can be improved by repacking the files using
the tools provided by the HDF5 libraries.

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>