EQUATIONS: Use wl, if the solution is uphysical.
If the energy function is positive for the lower bracket, it means that
the solution for energy function is unphysical (p<0). In this case use
the lower bracket which guarantees that the cell is physical.

The total energy should be corrected in this case, but this will be done
later.

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>