EQUATIONS: Use energy equation to calculate pressure.
In cons2prim_srmhd_adi() we use the energy equation to calculate the
pressure. If the enthalpy was calculated exactly, this the energy
equation gives pressure equal to the one obtained from the equation of
state. If the physical state couldn't be solved, the returned enthalpy
guarantees that the state is physical (v<1, p>0), but the solution
violates the equation of state.

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>