EQUATIONS: Fill the block borders in update_primitive_variables().
This should not be normally required, but it helps to avoid the
unexpected appearance of NaNs and the code execution termination due
to the floating point exceptions.

This should also help with more advanced open boundary conditions (such
as the conditions which keep the divergence or curl of vector fields
zero).

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>