BOUNDARIES: Update boundaries if any of block has update flag set.
We can have a situation when the whole block was changed and its update
flag was set. The boundaries of the block have to be updated too,
however, the neighbor's update flag is unset. This creates a undesired
situation of the block boundaries not being updated.

Just check if either the block or its neighbor has been set to update to
avoid the above situation.

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>