SCHEMES: Move reconstruction part to new subroutine.
The reconstruction in all three directions has been moved to a separate
subroutine in order to make it generic. It will allow for introducing a
full two or three dimensional reconstruction. This is for adiabatic
hydrodynamics so far, but other set of equations will be supported in
the following changes.

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>