INTERPOLATIONS: Implement multidimentional GP reconstruction.
This is truely multidimentional reconstruction based on Gaussian
Processes. All neighbors of the interpolation cell are taken into
account and properly weighted. The size of the neighbor region can be
controlled.

The monotonicity of the interpolated interfaces is detected and if the
case of lack it, the linear TVD reconstruction is applied. Therefore,
only the monotonic regions are interpolated using the high-order Gauss
Process.

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>