RECONNECTION: Rewrite the initial configuration of the magnetic field.
A new parameter 'zeta' was introduced to control the type of the
equilibrium. If zeta = 0.0 the total pressure equlilbrium is set, while
when zeta = 1.0 the force free magnetic equilibrium is set.

Also, instead of tanh(y) profile, the proper integral log(cosh(y)) is
used to define the discontinuous (reconnecting) magnetic field
component.

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>