SOURCES: Fix calculation of η J² term.
In the sum of the J² we used Fortran subroutine sum() and the sum was
done along the 2nd index. However, the interpretation of this index is
uncertain when the first array rank is 1.

Therefore, instead of using subroutine sum(), the calculation is done
directly.

After this change there are no strange effects appearing on the
boundaries of the blocks.

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>