INTERPOLATIONS: Slightly improve MGP reconstruction.
Subtract the cell value from the stencil and add it back after the
interpolated values once the interpolation is done. This significantly
decreases the interpolation errors in uniform areas where the variables
are different from zero.

Check the monotonicity by comparing to eps and not to zero. This does
not force the interpolation to go back to the TVD one if the difference
between the interpolated value and cell centered one is of the order of
numerical error.

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>