INTERPOLATIONS: Avoid infinite loop in interfaces_tvd() and interfaces_mgp().
In case an interpolated cell, which should be positive, is zero or negative,
and the sum of derivatives is zero as well, the correction of
derivatives can enter into an infinite loop. It shouldn't normally
happen, since such a situation is unphysical.

Avoid the infinite loop and print a warning about encountered problems.

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>