EQUATIONS: Improve unphysical cell correction.
If there are no enough physical neighbors, the cell cannot be corrected
and the code simply stops. Allow the code to continue the execution in
such a case by simply correcting the unphysical cells by replacing the
negative values of positive variables with a lower bounds, dmin and
pmin, for density and pressure, respectively.

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>